Created ability to have a station:parameter dict
Changed getData() function from accepting a list of stations and a single parameter (e.g. 'PREC') to where it takes in a list of dictionaries where each dictionary includes the station triplet and parameter (e.g. [{'302:OR:SNTL': 'PREC'}, {'653:OR:SNTL':'WTEQ'}.  This allows regression to occur with sensors at the same station as well as with sensors at other stations.